:bug: Check if there is dir public when cleaning

+++ clean.sh
@@ -1,7 +1,15 @@
 # reference from https://github.com/vibertthio/portfolio
 
 cd public/
-echo "clean all the html, js relevent files in ./public.."
+
+if [[ $PWD != /**/public ]]; then
+  echo "Dir public doesn't exist, create one and exit..."
+
+  mkdir public
+  exit 0
+fi
+
+echo "Clean all the html, js relevent files in ./public..."
 
 find . -name "*.scss.map" -type f -delete
 find . -name "*.scss" -type f -delete
@@ -11,7 +19,7 @@ find . -name "*.js.map" -type f -delete
 find . -name "*.js" -type f -delete
 cd ..
 
-echo "clean cache"
+echo "Clean cache..."
 rm -rf ./.cache/
 
 printf "\n"
